Turn bar revamp & character death visuals
Made the turn bar and characters representation much larger, this makes it much easier to see on a phone.
Disconnected the reliance on pixels to an abstract speed calculation that also allows easy balance to set how long a player would need to wait at most for a turn.
Resolves: #26 & #32

Added a grayscale shader to characters that has died to make it clearer for the player.